feat: 1 thread for gameplay and 1 thread for render

+++ Modules/Core/src/Scene/Camera.cpp
@@ -28,7 +28,7 @@ namespace BetterThanNothing
 		m_ZFar = zFar;
 	}
 
-	void Camera::Update(std::unique_ptr<Window>& window, f32 deltatime)
+	void Camera::Update(std::unique_ptr<Window>& window, double deltatime)
 	{
 		static bool firstMouse = true;
 		static f32 lastMouseX = 0;
