+++ Modules/Core/src/Scene/Camera.hpp
@@ -127,7 +127,7 @@ namespace BetterThanNothing
 		 * @brief Called every frame.
 		 * @param deltatime The time between the current frame and the last frame.
 		 */
-		void Update(std::unique_ptr<Window>& window, f32 deltatime);
+		void Update(std::unique_ptr<Window>& window, double deltatime);
 
 		/**
 		 * @brief Called when an Event is received.
